WIP: shell parser
